fix(admin-settings): space list after creating a space
Fixes the broken space list after creating a space due to the missing permissions.

Note that we currently can't fetch the permissions after creating a space because the server doesn't support `expand` on a drive GET request. Hence we fall back to a member count of 1 (always accurate directly after creating) and an empty list of manager names.