feat: proper resource fetch after upload
Instead of always reloading the entire file list after uploading resources, fetch them individually if the amount of uploaded files is <=5. This should alleviate some load on the server and provide a better UX, since the file list is not flickering when reloading.

Also, fixes the highlighting of new resources in the files list.